refactor: Arrow타입을 빼와서 파일로 생성, 메서드 은닉화 처리
