+++ BoxOffice/Controller/CollectionViewListCell.swift
@@ -77,23 +77,3 @@ final class CollectionViewListCell: UICollectionViewListCell {
         }
     }
 }
-
-extension CollectionViewListCell {
-    enum Arrow: String {
-        case upArrow = "▲"
-        case downArrow = "▼"
-        
-        var color: UIColor {
-            switch self {
-            case .upArrow:
-                return .red
-            case .downArrow:
-                return .blue
-            }
-        }
-        
-        init?(rawValue: Int) {
-            self = rawValue > 0 ? .upArrow : .downArrow
-        }
-    }
-}
